backend/requirements.txt
# Django & Extensions
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
django-celery-beat>=2.6,<3.0
django-celery-results>=2.5,<3.0
daphne>=4.1,<5.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Celery
celery>=5.4,<6.0
redis>=5.0,<6.0
flower>=2.0,<3.0

# LangChain & LangGraph
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# Langfuse
langfuse>=2.0

# Document Processing
pypdf>=4.0
python-docx>=1.0
unstructured>=0.15
tiktoken>=0.7
faiss-cpu>=1.8

# API & Serialization
pydantic>=2.0,<3.0

# Utilities
python-dotenv>=1.0
requests>=2.32
httpx>=0.27
aiohttp>=3.10
websockets>=12.0

# MCP
mcp>=1.0

# Monitoring
sentry-sdk>=2.0

# Testing
pytest>=8.0
pytest-django>=4.8
pytest-asyncio>=0.23
factory-boy>=3.3
